Add shortest path length function

+++ block.c
@@ -1,7 +1,7 @@
 #include "block.h"
 #include <stdlib.h>
 
-block *make_block(char **text, int lines) {
+block *make_block(char **text, size_t lines) {
   block *b = malloc(sizeof(block));
   b->text = text;
   b->lines = lines;
